+++ fm2prof/Fm2ProfRunner.py
@@ -18,7 +18,7 @@ from fm2prof import __version__
 from fm2prof.common import FM2ProfBase
 from fm2prof.CrossSection import CrossSection, CrossSectionHelpers
 from fm2prof.Export import Export1DModelData, OutputFiles
-from fm2prof.Import import FMDataImporter, FmModelData, ImportInputFiles
+from fm2prof.data_import import FMDataImporter, FmModelData, ImportInputFiles
 from fm2prof.IniFile import IniFile
 from fm2prof.MaskOutputFile import MaskOutputFile
 from fm2prof.RegionPolygonFile import RegionPolygonFile, SectionPolygonFile
@@ -228,7 +228,7 @@ class Fm2ProfRunner(FM2ProfBase):
 
         # Read FM model data
         fm2prof_fm_model_data = self._set_fm_model_data(map_file, css_file, regions, sections)
-        self.fm_model_data = FmModelData(fm2prof_fm_model_data)
+        self.fm_model_data = FmModelData(*fm2prof_fm_model_data)
 
         # Validate config file
         success: bool = self._validate_config_after_initalization()
