+++ tests/test_data_import.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from fm2prof.Import import FMDataImporter, FmModelData
+from fm2prof.data_import import FMDataImporter, FmModelData
 from tests.TestUtils import TestUtils, skipwhenexternalsmissing
 
 
@@ -10,9 +10,7 @@ class Test_FMDataImporter:
     @skipwhenexternalsmissing
     def test_when_map_file_without_czu_no_exception(self):
         # 1. Set up test data
-        test_map = Path(TestUtils.get_local_test_data_dir("main_test_data")).joinpath(
-            "fm_map.nc"
-        )
+        test_map = Path(TestUtils.get_local_test_data_dir("main_test_data")).joinpath("fm_map.nc")
         assert test_map.is_file()
 
         # 2. Set initial expectations
@@ -44,17 +42,13 @@ class Test_FmModelData:
         )
 
     @pytest.mark.parametrize("arg_list", [(""), (None)])
-    def test_when_argument_length_not_as_expected_then_exception_is_risen(
-        self, arg_list
-    ):
+    def test_when_argument_length_not_as_expected_then_exception_is_risen(self, arg_list):
         # 1. Set up test data
         arg_list = ["arg1", "arg2"]
 
         # 2. Set initial expectations
         expected_error_message = (
-            ""
-            + "Fm model data expects 5 arguments but only"
-            + " {} were given".format(len(arg_list))
+            "" + "Fm model data expects 5 arguments but only" + " {} were given".format(len(arg_list))
         )
 
         # 3. Run test
@@ -85,8 +79,8 @@ class Test_FmModelData:
         ]
         return_fm_model_data = None
 
-        # 2. Run test        
-        return_fm_model_data = FmModelData(arg_list)        
+        # 2. Run test
+        return_fm_model_data = FmModelData(arg_list)
 
         # 4. Verify final expectations
         assert return_fm_model_data is not None
@@ -120,9 +114,9 @@ class Test_FmModelData:
         # 2. Set expectations
         expected_css_data_list = [{dummy_key: 0}, {dummy_key: 1}]
 
-        #  3. Run test        
+        #  3. Run test
         return_fm_model_data = FmModelData(arg_list)
-        
+
         # 4. Verify final expectations
         assert return_fm_model_data is not None
         assert return_fm_model_data.css_data_list != css_data_dict
@@ -144,13 +138,11 @@ class Test_get_ordered_css_list:
 
         # 3. Run test
         return_list = FmModelData.get_ordered_css_list(test_dict)
-        
+
         # 4. Verify final expectations
         assert return_list is not None
         assert return_list == expected_list, (
-            ""
-            + "Expected return value {},".format(expected_list)
-            + " but return {} instead.".format(return_list)
+            "" + "Expected return value {},".format(expected_list) + " but return {} instead.".format(return_list)
         )
 
     @pytest.mark.parametrize("test_dict", [(""), (None), ({})])
@@ -161,13 +153,11 @@ class Test_get_ordered_css_list:
         # 2. Run test
         expected_list = []
 
-        # 3. Run test       
+        # 3. Run test
         return_list = FmModelData.get_ordered_css_list(test_dict)
-      
+
         # 4. Verify final expectations
         assert return_list is not None
         assert return_list == expected_list, (
-            ""
-            + "Expected return value {},".format(expected_list)
-            + " but return {} instead.".format(return_list)
+            "" + "Expected return value {},".format(expected_list) + " but return {} instead.".format(return_list)
         )
